+++ mth5/timeseries/spectre/multiple_station.py
